refactored slac package name to pyslac as 'slac' is not available to be used in pypi
