+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "slac"
+name = "pyslac"
 version = "0.8.0"
 description = "SLAC Protocol implementation, defined in ISO15118-3"
 authors = ["André Duarte <andre@switch-ev.com, andre14x@gmail.com>"]
