+++ pyslac/sockets/enums.py
