+++ pyslac/utils.py
