+++ tests/conftest.py
@@ -2,8 +2,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from slac.environment import Config
-from slac.session import SlacEvseSession
+from pyslac.environment import Config
+from pyslac.session import SlacEvseSession
 
 EVSE_ID = "DE*12*122333"
 IFACE = "en0"
@@ -21,8 +21,8 @@ def evse_mac():
 
 @pytest.fixture
 def evse_slac_session(dummy_config, evse_mac):
-    with patch("slac.session.get_if_hwaddr", new=Mock(return_value=evse_mac)):
-        with patch("slac.session.create_socket", new=Mock()):
+    with patch("pyslac.session.get_if_hwaddr", new=Mock(return_value=evse_mac)):
+        with patch("pyslac.session.create_socket", new=Mock()):
             evse_session = SlacEvseSession(EVSE_ID, IFACE, dummy_config)
             evse_session.reset_socket = Mock()
