--- tests/test_slac_version.py
@@ -1,5 +1,0 @@
-from slac import __version__
-
-
-def test_version():
-    assert __version__ == "0.1.0"
